refactor(monthly): decompose MonthlyAnalyzer into Facade + 3 Mixins
