+++ magma_cycling/analyzers/monthly/__init__.py
@@ -1,0 +1,1 @@
+"""MonthlyAnalyzer mixin modules for god class decomposition."""
